requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0

# Segmentation models
segmentation-models-pytorch>=0.3.3

# Data processing
albumentations>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Configuration
omegaconf>=2.3.0
pyyaml>=6.0

# Visualization & logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
tqdm>=4.65.0

# Optional: Wandb tracking
# wandb>=0.15.0

# Optional: Jupyter
# jupyter>=1.0.0
# ipywidgets>=8.0.0

# Download utility
requests>=2.31.0
